move test_error_message_on_startup
